Refactor and address review comments
