+++ cni/pkg/repair/repaircontroller_linux.go
@@ -57,18 +57,15 @@ func (c *Controller) repairPod(pod *corev1.Pod) error {
 	}
 	log = log.WithLabels("netns", netns)
 
+	redirector := redirectRunningPod
 	if c.cfg.NativeNftables {
-		if err := redirectRunningPodNFT(pod, netns); err != nil {
-			log.Errorf("failed to setup redirection: %v", err)
-			m.With(resultLabel.Value(resultFail)).Increment()
-			return err
-		}
-	} else {
-		if err := redirectRunningPod(pod, netns); err != nil {
-			log.Errorf("failed to setup redirection: %v", err)
-			m.With(resultLabel.Value(resultFail)).Increment()
-			return err
-		}
+		redirector = redirectRunningPodNFT
+	}
+
+	if err := redirector(pod, netns); err != nil {
+		log.Errorf("failed to setup redirection: %v", err)
+		m.With(resultLabel.Value(resultFail)).Increment()
+		return err
 	}
 
 	c.repairedPods[key] = pod.UID
