+++ tools/istio-nftables/pkg/capture/run_test.go
@@ -14,9 +14,7 @@
 package capture
 
 import (
-	"net/netip"
 	"path/filepath"
-	"reflect"
 	"testing"
 
 	"sigs.k8s.io/knftables"
@@ -295,62 +293,6 @@ func TestNftables(t *testing.T) {
 			// Collect the output/dump from all tables
 			dump := mock.Dump(tx)
 			compareToGolden(t, tt.name, dump)
-		})
-	}
-}
-
-func TestSeparateV4V6(t *testing.T) {
-	mkIPList := func(ips ...string) []netip.Prefix {
-		ret := []netip.Prefix{}
-		for _, ip := range ips {
-			ipp, err := netip.ParsePrefix(ip)
-			if err != nil {
-				panic(err.Error())
-			}
-			ret = append(ret, ipp)
-		}
-		return ret
-	}
-	cases := []struct {
-		name string
-		cidr string
-		v4   NetworkRange
-		v6   NetworkRange
-	}{
-		{
-			name: "wildcard",
-			cidr: "*",
-			v4:   NetworkRange{IsWildcard: true},
-			v6:   NetworkRange{IsWildcard: true},
-		},
-		{
-			name: "v4 only",
-			cidr: "10.0.0.0/8,172.16.0.0/16",
-			v4:   NetworkRange{CIDRs: mkIPList("10.0.0.0/8", "172.16.0.0/16")},
-		},
-		{
-			name: "v6 only",
-			cidr: "fd04:3e42:4a4e:3381::/64,ffff:ffff:ac10:ac10::/64",
-			v6:   NetworkRange{CIDRs: mkIPList("fd04:3e42:4a4e:3381::/64", "ffff:ffff:ac10:ac10::/64")},
-		},
-	}
-	for _, tt := range cases {
-		t.Run(tt.name, func(t *testing.T) {
-			cfg := constructTestConfig()
-			nftProvider := func(family knftables.Family, table string) (NftablesAPI, error) {
-				return NewMockNftables(family, table), nil
-			}
-			nftConfigurator, _ := NewNftablesConfigurator(cfg, nftProvider)
-			v4Range, v6Range, err := nftConfigurator.separateV4V6(tt.cidr)
-			if err != nil {
-				t.Fatal(err)
-			}
-			if !reflect.DeepEqual(v4Range, tt.v4) {
-				t.Fatalf("expected %v, got %v", tt.v4, v4Range)
-			}
-			if !reflect.DeepEqual(v6Range, tt.v6) {
-				t.Fatalf("expected %v, got %v", tt.v6, v6Range)
-			}
 		})
 	}
 }
